refactor(weather): Move components to feature-specific directory
- Relocated common components from lib/components/ to lib/features/weather/components/
to better organize feature-specific widgets.
- Updated import paths accordingly.
